Add file_selector_linux and file_selector_windows plugins

+++ mobile/ecommerce_app/lib/data/models/product_model.dart
@@ -12,7 +12,7 @@ class ProductModel extends ProductEntity {
         id: json['id'],
         name: json['name'],
         description: json['description'],
-        price: json['price'],
+        price: json['price'].toDouble(),
         imageUrl: json['imageUrl'],
       );
